refact(AuthorizedApps): move to sections folder
